Modified the EmailLists client scaffold and changed emails to singular, except for the view and action, and changed EmailsLists to EmailList. Implemented RESTful implementation of EmailLists successfully on client.
