+++ app/helpers/email_helper.rb
@@ -1,0 +1,2 @@
+module EmailHelper
+end
