--- app/helpers/emails_helper.rb
@@ -1,2 +1,0 @@
-module EmailsHelper
-end
