+++ app/models/email_list.rb
@@ -1,8 +1,8 @@
-class EmailsList < RestModel
+class EmailList < RestModel
 
 	# include_root_in_json :emails_list
 	# include_root_in_json false
 
-	has_many   :emails, :class_name => "Email"
+	has_many :email, :class_name => "Email"
 	belongs_to :app, :class_name => "App"
 end
