+++ app/views/email/list.html.erb
@@ -1,0 +1,16 @@
+<%= render 'layouts/navbar' %>
+
+<div class="container">
+	<h3><%= @app.name %></h3>
+	<ul class="list-group">
+		<% @list.each do |el| %>
+				<li class="list-group-item">
+					<a href="<%= root_url %>email/<%= el.id %>"><%= el.name %></a><hr>
+					<%= el.description %>
+				</li>
+		<% end %>
+	</ul>
+	
+	<b>Create New List</b>
+
+</div>
