--- app/views/emails/lists.html.erb
@@ -1,15 +1,0 @@
-<%= render 'layouts/navbar' %>
-
-
-<h3><%= @app.id %></h3>
-
-<div class="container">
-	<ul class="list-group">
-		<li class="list-group-item">Email List 1</li>
-		<li class="list-group-item">Email List 2</li>
-		<li class="list-group-item">Email List 3</li>
-	</ul>
-
-	<b>Create New List</b>
-	
-</div>
