+++ db/migrate/20161123061737_create_emails.rb
@@ -1,6 +1,6 @@
 class CreateEmails < ActiveRecord::Migration[5.0]
   def change
-    create_table :emails do |t|
+    create_table :email do |t|
 
       t.timestamps
     end
